refactor: correct Game styles
