+++ src/components/GameNotFound/GameNotFound.tsx
@@ -1,10 +1,9 @@
 import './GameNotFound.css';
 import React from 'react';
-import Button from '../../UI/Button/Button';
+import Button from '../UI/Button/Button';
 import { useTranslation } from 'react-i18next';
 
 function GameNotFound() {
-  
   const { t } = useTranslation();
 
   return (
