+++ src/components/GamePage/GamePage.css
@@ -27,4 +27,5 @@
   height: 100%;
   display: flex;
   justify-content: center;
+  align-items: center;
 }
