--- src/components/Games/Game1/Game1.css
@@ -1,24 +1,0 @@
-.game__word-label {
-  text-align: center;
-  display: block;
-  color: var(--rss-grey-dark-lighter);
-  margin: 0 0 10px 0;
-}
-
-.game__word-question {
-  text-align: center;
-  min-height: 40px ;
-  margin: 0 0 30px 0;
-}
-
-.game__word-answer {
-  text-align: center;
-  min-height: 40px ;
-  margin: 0 0 30px 0;
-}
-
-.game__group-controls {
-  display: flex;
-  flex-direction: row;
-  gap: 20px
-}
