--- src/components/Games/Game2/Game2.css
@@ -1,42 +1,0 @@
-.game {
-  padding: 80px 0;
-  display: flex;
-  flex-direction: column;
-  gap: 30px;
-}
-
-.game__wrapper {
-  height: 50px;
-  position: relative;
-}
-
-.game__word-row {
-  display: flex;
-  min-height: 50px;
-  flex-direction: row;
-  align-items: center;
-  justify-content: flex-start;
-  gap: 10px;
-}
-
-.game__char {
-  width: 50px;
-  height: 50px;
-  border: 1px solid #ccc;
-  border-radius: 10px;
-  background-color: #fff;
-  font-size: 25px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.game__wrapper .game__word-row {
-  position: absolute;
-}
-
-.game__char-button {
-  padding: 0;
-  width: 50px;
-  height: 50px;
-}
