--- src/components/Games/Game3/Game3.css
@@ -1,50 +1,0 @@
-.light .game {
-  --color-primary: var(--rss-blue);
-}
-
-.dark .game {
-  --color-primary: var(--rss-orange);
-}
-
-.game__container {
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.game__img-hear {
-  width: 120px;
-  height: 120px;
-  cursor: pointer;
-}
-
-.game__word {
-  text-align: center;
-}
-
-.game__main-traslation {
-  color: var(--color-primary);
-}
-
-.game__btn-next {
-  margin: 20px 0 0 0;
-}
-
-.game__listen-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.game__answer-container {
-  display: flex;
-  justify-content: space-around;
-  width: 600px;
-  padding-top: 60px;
-}
-
-.game__button-container {
-  width: 300px;
-  margin: 0 auto;
-}
