--- src/components/Games/Game4/Game4.css
@@ -1,63 +1,0 @@
-.light .game {
-  --color-bg: var(--rss-white);
-  --color-primary: var(--rss-blue);
-  --color-text: var(--rss-black);
-  --color-border: var(--rss-grey-light);
-}
-
-.dark .game {
-  --color-bg: var(--rss-grey-dark);
-  --color-primary: var(--rss-orange);
-  --color-text: var(--rss-grey-light);
-  --color-border: var(--rss-grey-light);
-}
-
-.game__container {
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.game__word {
-  text-align: center;
-}
-
-.game__main-traslation {
-  color: var(--color-primary);
-}
-
-.game__btn-next {
-  margin: 20px 0 0 0;
-}
-
-.game__input-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.game__input-submit {
-  display: flex;
-}
-
-.game__btn-rounded {
- border-radius: 0 8px 8px 0;
-  box-shadow: none;
-}
-
-.game__input-rounded {
-  display: block;
-  width: 100%;
-  min-width: 300px;
-  font-size: 16px;
-  font-weight: 400;
-  font-family: 'Inter', 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
-  border: 1px solid var(--color-border);
-  border-radius: 8px 0 0 8px;
-  padding: 16px;
-  background-color: var(--color-bg);
-  color: var(--color-text);
-  outline: none;
-  transition: all 0.1s linear;
-}
