--- src/components/Games/Games.css
@@ -1,6 +1,0 @@
-.games__container {
-  padding: 20px 0;
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  grid-gap: 10px;
-}
